+++ geo/CHANGES.md
@@ -5,6 +5,11 @@
 - Fix `CoordinatePosition` for `LineString` to handle dimensionally collapsed input  e.g. `LINESTRING(0 0)` is treated like `POINT(0 0)`.
   - <https://github.com/georust/geo/pull/1483>
 - Fix `CoordinatePosition` for `Triangle` to correctly return `CoordPos::OnBoundary` for coordinate within vertical segment.
+- Add `MonotoneChain`-backed geometry types
+  - Variant of `LineString`, `MultiLineString`, `Polygon`, `MultiPolygon` which are backed by Monotone Chains
+  - Preprocessing cost to construct these types, but provide a significant performance boost for intersects and contains_properly checks
+  - <https://github.com/georust/geo/issues/1466>
+  - <https://github.com/georust/geo/pull/1467>
 
 ## 0.32.0 - 2025-12-05